+++ main/npm/registry.ts
